Move transpilation of literals into separate file
